2010-01-14 04:04 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/binnumx.c   * contrib/xpp/binnumx.c   * contrib/hbmisc/stringsx.c   * contrib/hbtip/utils.c   * contrib/hbcairo/context.c   * contrib/hbcairo/core.c     * ULONG -> HB_SIZE
  * contrib/hbgt/ascposgt.c
    * ULONG -> HB_SIZE
    ! hb_parni() -> hb_parnl()

  * contrib/hbsqlit3/hbsqlit3.c
    ! ULONG -> int (var used in 3rd party interface)

  * contrib/hbmzip/hbmzip.c
    * Formatting.

  * contrib/hbclipsm/num.c
    % hb_retnl() -> hb_retni()
    * ULONG -> HB_SIZE

  * contrib/hbgd/gdwrp.c
    % Deleted few unnecessary ULONG casts.
    ! Fixed win64 problem where file handle was read from hb_parnl()
      instead of hb_parnint().
    ! Fixed 'hb_parnl( 2 ) > -1' code checking for valid file handle
      to 'hb_parnint( 2 ) != FS_ERROR'.
    * ULONG -> HB_SIZE
    % Elmininated few variables.
    * Formatting.

  * contrib/hbmisc/strfmt.c
    * ULONG -> int, HB_SIZE
    * BOOL -> HB_BOOL.

+++ harbour/contrib/hbsqlit3/hbsqlit3.c
@@ -753,11 +753,11 @@ HB_FUNC( SQLITE3_PREPARE )
       if( SQL )
       {
          const char     *pSQL = hb_itemGetCPtr( SQL );
-         ULONG          ulLen = hb_itemGetCLen( SQL );
+         int            ulLen = ( int ) hb_itemGetCLen( SQL );
          psqlite3_stmt  pStmt;
          const char     *pszTail;
 
-         if( sqlite3_prepare_v2(pHbSqlite3->db, pSQL, ulLen, &pStmt, &pszTail) == SQLITE_OK )
+         if( sqlite3_prepare_v2( pHbSqlite3->db, pSQL, ulLen, &pStmt, &pszTail ) == SQLITE_OK )
          {
             hb_retptr( pStmt );
          }
